+++ R/exceedance.R
@@ -257,43 +257,21 @@ exceedance <-
     thresh <- int_mean <- int_max <- int_cum <- exceedance_rel_thresh <-
       int_mean_abs <- int_max_abs <- int_cum_abs <- temp <- NULL ###
 
-    ## This code to replace the plyr bits, below. RWS to first find the problem with int_mean,
-    ## int_max, int_var, int_cum, int_mean_abs, int_max_abs, int_var_abs and int_cum_abs being
-    ## incorrectly reported (all show identical values...)
-    # exceedances <- cbind(exceedances,
-    #                      exceedances_list %>%
-    #                        dplyr::bind_rows(.id = "exceedance_no") %>%
-    #                        dplyr::group_by(exceedance_no) %>%
-    #                        dplyr::summarise(date_peak = date[temp == max(temp)][1],
-    #                                         int_mean = mean(exceedance_rel_thresh),
-    #                                         int_max = max(exceedance_rel_thresh),
-    #                                         int_var = sqrt(stats::var(exceedance_rel_thresh)),
-    #                                         int_cum = max(cumsum(exceedance_rel_thresh)),
-    #                                         int_mean_abs = mean(temp),
-    #                                         int_max_abs = max(temp),
-    #                                         int_var_abs = sqrt(stats::var(temp)),
-    #                                         int_cum_abs = max(cumsum(temp))) %>%
-    #                        dplyr::arrange(as.numeric(exceedance_no)) %>%
-    #                        dplyr::select(-exceedance_no))
-
-    exceedances$date_peak <-
-      plyr::ldply(exceedances_list, function(x) x$date[x$temp == max(x$temp)][1])[, 2]
-    exceedances$int_mean <-
-      plyr::ldply(exceedances_list, function(x) mean(x$exceedance_rel_thresh))[, 2]
-    exceedances$int_max <-
-      plyr::ldply(exceedances_list, function(x) max(x$exceedance_rel_thresh))[, 2]
-    exceedances$int_var <-
-      plyr::ldply(exceedances_list, function(x) sqrt(stats::var(x$exceedance_rel_thresh)))[, 2]
-    exceedances$int_cum <-
-      plyr::ldply(exceedances_list, function(x) max(cumsum(x$exceedance_rel_thresh)))[, 2]
-    exceedances$int_mean_abs <-
-      plyr::ldply(exceedances_list, function(x) mean(x$temp))[, 2]
-    exceedances$int_max_abs <-
-      plyr::ldply(exceedances_list, function(x) max(x$temp))[, 2]
-    exceedances$int_var_abs <-
-      plyr::ldply(exceedances_list, function(x) sqrt(stats::var(x$temp)))[, 2]
-    exceedances$int_cum_abs <-
-      plyr::ldply(exceedances_list, function(x) max(cumsum(x$temp)))[, 2]
+    exceedances <- cbind(exceedances,
+                         exceedances_list %>%
+                           dplyr::bind_rows(.id = "exceedance_no") %>%
+                           dplyr::group_by(exceedance_no) %>%
+                           dplyr::summarise(date_peak = date[temp == max(temp)][1],
+                                            int_mean = mean(exceedance_rel_thresh),
+                                            int_max = max(exceedance_rel_thresh),
+                                            int_var = sqrt(stats::var(exceedance_rel_thresh)),
+                                            int_cum = max(cumsum(exceedance_rel_thresh)),
+                                            int_mean_abs = mean(temp),
+                                            int_max_abs = max(temp),
+                                            int_var_abs = sqrt(stats::var(temp)),
+                                            int_cum_abs = max(cumsum(temp))) %>%
+                           dplyr::arrange(as.numeric(exceedance_no)) %>%
+                           dplyr::select(-exceedance_no))
 
     exceedance_rel_thresh <- t_series$temp - t_series$thresh
     A <- exceedance_rel_thresh[exceedances$index_start]
@@ -311,19 +289,19 @@ exceedance <-
         difftime(exceedances$date_peak, exceedances$date_start, units = "days")) + 0.5),
       NA
     )
-    
+
     D <- exceedance_rel_thresh[exceedances$index_stop]
     E <- t_series$temp[exceedances$index_stop + 1]
     F <- t_series$thresh[exceedances$index_stop + 1]
     exceedance_rel_thresh_end <- 0.5 * (D + E - F)
-    
+
     exceedances$rate_decline <- ifelse(
       exceedances$index_stop < nrow(t_series),
       (exceedances$int_max - exceedance_rel_thresh_end) / (as.numeric(
         difftime(exceedances$date_stop, exceedances$date_peak, units = "days")) + 0.5),
       NA
     )
-    
+
     if (below) {
       exceedances <- exceedances %>% dplyr::mutate(
         int_mean = -int_mean,