+++ NEWS.md
@@ -6,10 +6,11 @@ output: pdf_document
 ## Changes in version 0.15.9 (19 June 2017)
 * Warnings given by exceedance() when exceedances begin/end on the first/last day of the
   time series have been corrected --- it will now report an NA when the exceedance
-  occurs at the start/end of the time series. 
+  occurs at the start/end of the time series.
+* Some plyr routines phased out in exceedance function in favour of a dplyr approach.
   
 ## Changes in version 0.15.8 (18 June 2017)
-* Further remove need for plyr functions --- use dplyr instead.
+* Further removed need for plyr functions --- use dplyr instead.
 
 ## Changes in version 0.15.7 (18 June 2017)
 * Warnings given by detect() when events begin/end on the first/last day of the